% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/castform-package.R
\docType{package}
\name{castform}
\alias{castform-package}
\alias{castform}
\title{castform: Download Environment Canada Weather Station Data}
\description{
Downloads hourly weather station data (as csv files) from Environment Canada.
}
\section{Main functions}{

\itemize{
  \item \code{\link{get_single_station_file}}: The primary function to download one hourly data file.
  \item \code{\link{get_multiple_station_files}}: Downloads multiple hourly data files.
  \item \code{\link{province_station_files}}: Downloads data files by province.
  \item \code{\link{year_range_station_files}}: Downloads data files by year range.
}
}

\author{
\strong{Maintainer}: Shawn Yip \email{shawniceyip@trentu.ca}

}
\keyword{internal}
