% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_missingness_function.R
\name{data_missingness_table}
\alias{data_missingness_table}
\title{Data Missingness Table}
\usage{
data_missingness_table(
  db_name = NULL,
  db_dir = "station_data",
  output_dir = "station_data"
)
}
\arguments{
\item{db_name}{Character: The name of the database}

\item{db_dir}{The directory of the database, If left unchanged, will default to package's default created directory "station_data".}

\item{output_dir}{The created download folder and file path. If left unchanged, will create a new "station_data" folder in the working directory.}
}
\description{
Creates an `.html` output table with data missingness from each station within a database. Stores the expected, account, and percent missing counts for each variable.
}
