% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.climate_regions_stations}
\alias{.climate_regions_stations}
\title{Identify Station Climate Region}
\usage{
.climate_regions_stations(stationID = NULL)
}
\arguments{
\item{stationID}{Station ID from queried results from `heatwave_detector()`}
}
\value{
If a matching Station ID is found, will return the daytime maximum and night time minimum temperature threshold for that station as a list.
}
\description{
Match the station to it's climate region and extreme temperature threshold. Last updated April 17, 2026.
}
