% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.mapping}
\alias{.mapping}
\title{Create station map}
\usage{
.mapping(input_df = NULL, file_path = file_path)
}
\arguments{
\item{input_df}{Provided dataframe}

\item{file_path}{Provided filepath}
}
\value{
A `.png` of selected stations
}
\description{
Allows for the visualization of all the stations in the metadata or selected stations.
}
