% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_files_function.R
\name{get_all_files}
\alias{get_all_files}
\title{Get All Hourly Station Files}
\usage{
get_all_files(root_folder = "station_data")
}
\arguments{
\item{root_folder}{The created download folder and file path. If left unchanged, will create a new "station_data" folder in the working directory.}
}
\description{
Downloads all Environment Canada data from all stations with hourly data available.
}
