% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata_function.R
\name{get_metadata}
\alias{get_metadata}
\title{Download Station Metadata}
\usage{
get_metadata()
}
\description{
Download the latest station inventory list from Environment Canada. This will download the latest .csv file and load it into your R environment as an .rda file. This needs to be loaded in FIRST to ensure the other functions have data to work with.
}
