% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_multiple_stations_function.R
\name{get_multiple_station_files}
\alias{get_multiple_station_files}
\title{Get Multiple Station Files}
\usage{
get_multiple_station_files(
  station_name = NULL,
  station_id = NULL,
  number_of_files = NULL,
  year = NULL,
  month = NULL,
  parallel_threshold = 50
)
}
\arguments{
\item{station_name}{Character: The name of the weather station of interest.}

\item{station_id}{Numeric Integer: The unique station ID of the weather station of interest.}

\item{number_of_files}{Numeric Integer: The number of files the user wishes to download.}

\item{year}{Numeric Integer: The year of the data pull. If left empty, will default to the first year for data collection for that particular station.}

\item{month}{Numeric Integer: The month of the data pull (1 - 12). If left empty, will default to January (1).}

\item{parallel_threshold}{Numeric Integer: The required number of files to trigger parallel downloads. If left unchanged, parallelization will occur for downloads of 50 files or more.}
}
\description{
Downloads a multiple weather station files from Environment Canada
}
