% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_single_station_function.R
\name{get_single_station_file}
\alias{get_single_station_file}
\title{Get Single Station File}
\usage{
get_single_station_file(
  station_name = NULL,
  station_id = NULL,
  year = NULL,
  month = NULL,
  out_dir = "station_data",
  HLY_station_info = NULL
)
}
\arguments{
\item{station_name}{Character: The name of the weather station of interest.}

\item{station_id}{Numeric Integer: The unique station ID of the weather station of interest.}

\item{year}{Numeric Integer: The year of the data pull. If left empty, will default to the first year for data collection for that particular station.}

\item{month}{Numeric Integer: The month of the data pull (1 - 12). If left empty, will default to January (1).}

\item{out_dir}{The created download folder and file path. If left unchanged, will create a new "station_data" folder in the working directory.}

\item{HLY_station_info}{Dataframe: Station metadata}
}
\description{
Downloads a single weather station file from Environment Canada
}
