% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_files_function.R
\name{get_station_files}
\alias{get_station_files}
\title{Get Station Files}
\usage{
get_station_files(
  station_name = NULL,
  station_id = NULL,
  year = NULL,
  month = NULL,
  parallel_threshold = 50,
  root_folder = "station_data",
  HLY_station_info = NULL
)
}
\arguments{
\item{station_name}{Character: The name of the weather station of interest.}

\item{station_id}{Numeric Integer: The unique station ID of the weather station of interest.}

\item{year}{Numeric Integer: The year of the data pull. If left empty, will default to downloading all years with hourly data for that particular station.}

\item{month}{Numeric Integer: The month of the data pull (1 - 12). If left empty, will default to downloading data for all months (1-12).}

\item{parallel_threshold}{Numeric Integer: The required number of files to trigger parallel downloads. If left unchanged, parallelization will occur for downloads of 50 files or more.}

\item{root_folder}{The created download folder and file path. If left unchanged, will create a new "station_data" folder in the working directory.}

\item{HLY_station_info}{Dataframe: Station metadata}
}
\description{
Download Environment Canada weather data by station
}
