% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_means_plot.R
\name{plot_yearly_means}
\alias{plot_yearly_means}
\title{Yearly Mean Plots}
\usage{
plot_yearly_means(
  db_name = NULL,
  db_dir = "station_data",
  output_dir = "station_data",
  output_name = NULL,
  write_csv = FALSE
)
}
\arguments{
\item{db_name}{Character: The name of the database}

\item{db_dir}{Character: The directory of the database, If left unchanged, will default to package's default created directory "station_data".}

\item{output_dir}{Character: The created download folder and file path. If left unchanged, will create a new "station_data" folder in the working directory.}

\item{output_name}{Character: The name of the output file. If left unfilled, the function will name the file "db_name_missingness_table.html"}

\item{write_csv}{Logical: If TRUE prints a csv copy of the results}
}
\value{
An `.html` output line plot visualizing the data.
}
\description{
Summarizes the average of each variable over time.
}
