% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_map_function.R
\name{station_map}
\alias{station_map}
\title{Plot Stations on Map}
\usage{
station_map(
  db_name = NULL,
  db_dir = "station_data",
  output_dir = "station_data",
  output_name = NULL,
  metadata_stations = FALSE
)
}
\arguments{
\item{db_name}{Character: The name of the database}

\item{db_dir}{Character: The directory of the database, If left unchanged, will default to package's default created directory "station_data".}

\item{output_dir}{Character: The created download folder and file path. If left unchanged, will create a new "station_data" folder in the working directory.}

\item{output_name}{Character: The chosen name of the output `png` object.}

\item{metadata_stations}{Logical: Plot all stations in the metadata (default = FALSE)}
}
\value{
A `.png` of the created map
}
\description{
Allows for users to plot all metadata stations or selected stations on a map of Canada.
}
