% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_validation_function.R
\name{validate_database}
\alias{validate_database}
\title{Validate Created Database}
\usage{
validate_database(db_name = NULL, db_dir = "station_data")
}
\arguments{
\item{db_name}{Character: The name of the database}

\item{db_dir}{The directory of the database, If left unchanged, will default to "station_data".}
}
\description{
Used to validate databases created by `build_station_database`. This function will check for tables created within the database, count the number of records within each table, and print the first five observation records.
}
