% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_range_function.R
\name{year_range_station_files}
\alias{year_range_station_files}
\title{Download Stations by Year Range}
\usage{
year_range_station_files(
  station_name = NULL,
  station_id = NULL,
  start_year = NULL,
  end_year = NULL,
  parallel_threshold = 50,
  out_dir = "station_data",
  HLY_station_info = NULL
)
}
\arguments{
\item{station_name}{Character: The name of the weather station of interest.}

\item{station_id}{Numeric Integer: The unique station ID of the weather station of interest.}

\item{start_year}{Numeric Integer: The start year of the data pull. If left empty, will default to the first year for data collection for that particular station.}

\item{end_year}{Numeric Integer: The end year of the data pull. If left empty, it will default to only downloading one year of data from the start_year}

\item{parallel_threshold}{Numeric Integer: The required number of files to trigger parallel downloads. If left unchanged, parallelization will occur for downloads of 50 files or more.}

\item{HLY_station_info}{Dataframe: Station metadata}

\item{root_folder}{The created download folder and file path. If left unchanged, will create a new "station_data" folder in the working directory.}
}
\description{
Downloads a multiple weather station files collected between a specified time period from Environment Canada
}
