% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/province_stations_function.R
\name{province_station_files}
\alias{province_station_files}
\title{Province Station Files}
\usage{
province_station_files(
  province = NULL,
  year = NULL,
  month = NULL,
  parallel_threshold = 50,
  root_folder = "station_data"
)
}
\arguments{
\item{province}{Character: The province or territory of interest.}

\item{year}{Numeric Integer: The year of the data pull. If left empty, will default to the first year for data collection for that particular station.}

\item{month}{Numeric Integer: The month of the data pull (1 - 12). If left empty, will default to January (1).}

\item{parallel_threshold}{Numeric Integer: The required number of files to trigger parallel downloads. If left unchanged, parallelization will occur for downloads of 50 files or more.}

\item{root_folder}{The created download folder and file path. If left unchanged, will create a new "station_data" folder in the working directory.}
}
\description{
Download Environment Canada weather station files by province.
}
