% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_strings_function.R
\name{pull_missing_strings}
\alias{pull_missing_strings}
\title{Pull Strings of Missing Data}
\usage{
pull_missing_strings(
  db_name = NULL,
  db_dir = "station_data",
  output_dir = "station_data",
  output_name = NULL,
  write_csv = FALSE
)
}
\arguments{
\item{db_name}{Character: The name of the database}

\item{db_dir}{Character: The directory of the database, If left unchanged, will default to package's default created directory "station_data".}

\item{output_dir}{Character: The created download folder and file path. If left unchanged, will create a new "station_data" folder in the working directory.}

\item{output_name}{Character: The name of the output file. If left unfilled, the function will name the file "db_name_missingness_table.html"}
}
\value{
A `.html` output table and plot that displays the length of the data gap (in hours) as well as the start and end date/time.
}
\description{
Identifies when data is missing from the database.
}
